회원탈퇴

+++ nnb/src/main/java/ezen/nnb/member/service/MyInfoServiceImpl.java
@@ -5,6 +5,7 @@ import javax.annotation.Resource;
 import org.apache.log4j.Logger;
 import org.springframework.stereotype.Service;
 
+import ezen.nnb.member.dao.LoginDAO;
 import ezen.nnb.member.dao.MyInfoDAO;
 
 @Service("myInfoService")
@@ -13,6 +14,9 @@ public class MyInfoServiceImpl implements MyInfoService{
 
 	@Resource(name="myInfoDAO")
 	private MyInfoDAO myInfoDAO;
+	
+	@Resource(name="loginDAO")
+	private LoginDAO loginDAO;
 	
 	@Override
 	public Map<String, Object> selectMyInfoDetail(Map<String, Object> map) throws Exception {
@@ -23,8 +27,16 @@ public class MyInfoServiceImpl implements MyInfoService{
 		myInfoDAO.updateMyInfoModify(map);
 	}
 	@Override
-	public void deleteMyInfo(Map<String, Object> map) throws Exception {
-		myInfoDAO.deleteMyInfo(map);
+	public int deleteMyInfo(Map<String, Object> map) throws Exception {
+		int res;
+		Map<String,Object> member = loginDAO.selectId(map);
+		if(member.get("MEM_PW").equals(map.get("MEM_PW"))){
+			myInfoDAO.deleteMyInfo(map);
+			res = 1;
+		}else {
+			res = 0;
+		}
+		return res;
 	}
 		
 }
